add userOpsControllers directory with atomic controllers for async requests
